Update RP2040_Central_BLE_Node.ino
This update improves raw IMU data transmitted from an Arduino Nano 33 BLE acting as the peripheral to an Arduino Nano RP2040 Connect acting as the central node. It adds support for quaternion estimation of position through Madgwick filter. This is specifically the central node and prints this data to a serial monitor.