Update Nano_33_BLE_IMU_Peripheral_Node.ino
This update improves raw IMU data transmitted from an Arduino Nano 33 BLE acting as the peripheral to an Arduino Nano RP2040 Connect acting as the central node. It adds support for quaternion estimation of position through the Madgwick filter. This is specifically the peripheral node and sends the data to the central node or to the micro-ROS central node. THIS IS THE PERIPHERAL SCRIPT AND IS TO BE USED WITH BOTH THE MICRO-ROS PUBLISHER AND THE STANDARD ARDUINO SERIAL PUBLISHER.